moved equations to new above_ground file
